+++ src/main/java/mods/belgabor/bitdrawers/config/ConfigManager.java
@@ -14,6 +14,8 @@ public class ConfigManager {
     
     public boolean debugTrace = false;
     public int bitdrawerStorage = 16;
+    public boolean allowBagMultiInsertion = true;
+    public boolean allowChiseledBlockMultiInsertion = true;
     
     public ConfigManager(File config) {
         this.config = new Configuration(config);
@@ -28,7 +30,15 @@ public class ConfigManager {
         bitdrawerStorage = config.get(Configuration.CATEGORY_GENERAL, "bitdrawerBaseStorage", 16,
                 "Base storage of a bit drawer (stacks).").setRequiresWorldRestart(true)
                 .setLanguageKey(LANG_PREFIX + "prop.bitdrawerBaseStorage").getInt();
-        
+
+        allowBagMultiInsertion = config.get(Configuration.CATEGORY_GENERAL, "allowBagMultiInsertion", true,
+                "If set the contents of all bags in a players inventory will try to be inserted on a double right-click on the bit drawer controller.")
+                .setLanguageKey(LANG_PREFIX + "prop.allowBagMultiInsertion").getBoolean();
+
+        allowChiseledBlockMultiInsertion = config.get(Configuration.CATEGORY_GENERAL, "allowChiseledBlockMultiInsertion", true,
+                "If set all Chisels & Bits blocks in a players inventory will try to be inserted on a double right-click on the bit drawer controller.")
+                .setLanguageKey(LANG_PREFIX + "prop.allowChiseledBlockMultiInsertion").getBoolean();
+
         if (config.hasChanged())
             config.save();
     }